ci: Build and push skydive-flow-exporter Docker image
This modification to the create-docker-image CI scripts is intended to
create another image besides the official Skydive image (which is used
to run agents and analyzers): an image that contains the `allinone` flow
exporter (from contrib/exporters/allinone).

The new images are called `skydive/skydive-flow-exporter` (for the
official and latest releases) and
`skydive/skydive-flow-exporter-snapshots` (for snapshot releases).

This will allow users of the Skydive flow exporters to simply pull an
image from the Skydive repository directly into their Kubernetes
deployment.