exporters: Add Context field to resulting flow JSON
Instead of packing information in the `Network.A_Name` string field, we
introduce the following output format:

    {
      "Network": {
        "A": "172.17.0.13",
        "B": "8.8.8.8"
        "A_Name": "0_0_my-namespace/my-pod-name_0",
      },
      "Context": {
        "A": {
          "Type": "pod",
          "Name": "my-namespace/my-pod-name",
          "Set": "ReplicaSet:my-namespace/my-replicaset"
        }
      }

This allows for easier and safer addition of more enrichment information
from the Skydive topology into the reported flows.

We add an IPToContext function to the resolves, which replaces the
IPToName functionality.  Context is stored in a local cache per IP+TID.

Repeated Gremlin queries were extracted to gremlin_queries.go.

Note that for now the A_Name field is still kept as-is for backward
compatibility.