contrib/exporters: Don't fill A_Name/B_Name when can't find name for IP address
Following the addition of resolvers, secadvisor exporter began filling
the `Network.A_Name` and `Network.B_Name` fields of the resulting flow
with the original IP address (`Network.A` or `Network.B`) in case the
name resolving fails.

Besides including redundant information in the flow JSON, this also
breaks the current protocol with Security Advisor's data controller,
which expects the `A_Name` to be either empty/undefined or a string with
the specific format of `0_0_container-name_0`.

This fix simply puts an empty string in the name, which later means the
field will be omitted when the flow is encoded as JSON.